feat(refactor) - fixes typo in models - updates validation and middleware
