+++ server/middleware/validation/object-id-validation.js
@@ -1,0 +1,18 @@
+import mongoose from 'mongoose'
+import { check } from 'express-validator'
+
+const objectIdValidation = (idName) => {
+  return [
+    check(idName)
+      .trim()
+      .escape()
+      .notEmpty()
+      .withMessage(`${idName} can't be empty.`)
+      .custom((id) => mongoose.Types.ObjectId.isValid(id))
+      .withMessage(`Invalid ${idName} format.`)
+      .bail() // stops execution of current chain if there's an error
+      .customSanitizer((id) => new mongoose.Types.ObjectId(String(id))),
+  ]
+}
+
+export default objectIdValidation
