+++ server/models/Plan.js
@@ -35,6 +35,6 @@ const planSchema = new mongoose.Schema(
   }
 )
 
-const Plan = mongoose.model.Plan || mongoose.model('Plan', planSchema)
+const Plan = mongoose.models.Plan || mongoose.model('Plan', planSchema)
 
 export default Plan
