+++ server/models/User.js
@@ -42,6 +42,6 @@ userSchema.pre('save', async function (next) {
   next()
 })
 
-const User = mongoose.model('User', userSchema)
+const User = mongoose.models.User || mongoose.model('User', userSchema)
 
 export default User
